backend/voice_agent_requirements.txt
livekit-agents[deepgram,cartesia,anthropic,silero,turn-detector]~=1.0
python-dotenv~=1.0
